requirements.txt
Flask
markupsafe
openai==1.3.4
python-dotenv 
tiktoken

